here we go again, more fixes this time edition

+++ oofsyDoofsy.py
@@ -18,8 +18,8 @@ import psutil
 print("Checking for Roblox installation...")
 time.sleep(0.5)
 if not os.path.exists(os.environ['LOCALAPPDATA'] + '/Roblox'):
-    print("Oops! You dont see to have Roblox installed! Without it, this script cannot continue!")
-    time.sleep(0.7)
+    print("Oops! You dont seem to have Roblox installed! Without it, this script cannot continue!")
+    time.sleep(1)
     exit(1)
 else:
     print("Found installation!")
@@ -41,15 +41,15 @@ if backup == "Y" or backup == "y" or backup == "":
     os.system("cls")
     print("Now, choose the file you want to use as the new 'ouch.ogg' file.")
     oggcheck = fd.askopenfilename(defaultextension=".ogg", filetypes=[("Ogg files", "*.ogg")], title="Choose a file.", initialdir="./")
-    os.rename(oggcheck, directory + 'ouch.ogg')
     if not oggcheck.endswith(".ogg"):
         print("Oops! The file you selected is not an .ogg file! Please try again.")
         time.sleep(0.7)
         exit(1)
     else:
         print("Done! The new 'ouch.ogg' file has been selected! Renaming...")
-        os.remove(directory + 'ouch.ogg')
-        os.rename(oggcheck, directory + 'ouch.ogg')
+        if os.path.exists(directory + 'ouch.ogg'):
+            os.remove(directory + 'ouch.ogg')
+        os.rename(os.path.basename(oggcheck), 'ouch.ogg')
         shutil.move('ouch.ogg', directory + 'ouch.ogg')
         time.sleep(1)
         os.system("cls")
@@ -57,17 +57,12 @@ if backup == "Y" or backup == "y" or backup == "":
             print("Roblox is currently running. Please close it before continuing.")
             time.sleep(0.7)
             exit(1)
-        else:
-            print("Everything is done! You can now open Roblox and enjoy your new modified file!")
-            time.sleep(0.7)
-            exit(0)
 elif backup == "N" or backup == "n":
     print("Backup is STRONGLY recommended, but continuing without backup...")
     time.sleep(1)
     os.system("cls")
     print("Now, choose the file you want to use as the new 'ouch.ogg' file.")
     oggcheck = fd.askopenfilename(defaultextension=".ogg", filetypes=[("Ogg files", "*.ogg")], title="Choose a file.", initialdir="./")
-    #oggcheck = os.path.basename(oggcheck)
     for directory in glob.glob(os.environ['LOCALAPPDATA'] + '/Roblox/Versions/version-*/content/sounds/'):
         if os.path.exists(directory + 'ouch.ogg'):
             os.remove(directory + 'ouch.ogg')
